scripts/requirements.txt
openai
supabase
python-dotenv
langchain
PyPDF2
requests